[refactor] deliverer 4계층 리펙토링-1
