+++ deliverer/src/main/java/com/sparta/rooibos/deliverer/application/aop/RoleCheckAspect.java
@@ -1,6 +1,6 @@
-package com.sparta.rooibos.deliverer.infrastructure.aop;
+package com.sparta.rooibos.deliverer.application.aop;
 
-import com.sparta.rooibos.deliverer.infrastructure.auditing.UserAuditorContext;
+import com.sparta.rooibos.deliverer.application.auditing.UserAuditorContext;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.aspectj.lang.ProceedingJoinPoint;
