+++ deliverer/src/main/java/com/sparta/rooibos/deliverer/application/auditing/UserAuditorContext.java
@@ -1,4 +1,4 @@
-package com.sparta.rooibos.deliverer.infrastructure.auditing;
+package com.sparta.rooibos.deliverer.application.auditing;
 
 import java.util.Optional;
