+++ deliverer/src/main/java/com/sparta/rooibos/deliverer/application/feign/HubService.java
@@ -1,0 +1,8 @@
+package com.sparta.rooibos.deliverer.application.feign;
+
+import java.util.UUID;
+
+public interface HubService {
+    UUID getHubIdByEmail(String email);
+    boolean checkHub(UUID hubId);
+}
