+++ deliverer/src/main/java/com/sparta/rooibos/deliverer/application/feign/UserService.java
@@ -1,0 +1,9 @@
+package com.sparta.rooibos.deliverer.application.feign;
+
+import com.sparta.rooibos.deliverer.infrastructure.feign.dto.UserResponse;
+
+import java.util.UUID;
+
+public interface UserService {
+    UserResponse getUserByMaster(UUID userId);
+}
