+++ deliverer/src/main/java/com/sparta/rooibos/deliverer/application/service/port/DelivererService.java
@@ -4,7 +4,6 @@ import com.sparta.rooibos.deliverer.application.dto.request.DelivererRequest;
 import com.sparta.rooibos.deliverer.application.dto.request.DelivererSearchRequest;
 import com.sparta.rooibos.deliverer.application.dto.response.DelivererListResponse;
 import com.sparta.rooibos.deliverer.application.dto.response.DelivererResponse;
-import com.sparta.rooibos.deliverer.domain.entity.DelivererType;
 
 import java.util.UUID;
 
@@ -14,6 +13,6 @@ public interface DelivererService {
     void deleteDeliverer(UUID delivererId);
     DelivererResponse getDeliverer(UUID delivererId);
     DelivererListResponse searchDeliverers(DelivererSearchRequest request);
-    DelivererResponse assignNextDeliverer(UUID hubId, DelivererType type);
+    DelivererResponse assignNextDeliverer(UUID hubId, String type);
     void cancelAssignment(UUID delivererId);
 }
