+++ deliverer/src/main/java/com/sparta/rooibos/deliverer/infrastructure/auditing/AuditorAwareImpl.java
@@ -1,5 +1,6 @@
 package com.sparta.rooibos.deliverer.infrastructure.auditing;
 
+import com.sparta.rooibos.deliverer.application.auditing.UserAuditorContext;
 import org.springframework.data.domain.AuditorAware;
 import org.springframework.stereotype.Component;
