+++ deliverer/src/main/java/com/sparta/rooibos/deliverer/infrastructure/feign/UserClient.java
@@ -1,5 +1,6 @@
 package com.sparta.rooibos.deliverer.infrastructure.feign;
 
+import com.sparta.rooibos.deliverer.application.feign.UserService;
 import com.sparta.rooibos.deliverer.infrastructure.feign.dto.UserResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -8,8 +9,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import java.util.UUID;
 
 @FeignClient(name = "user-service")
-public interface UserClient {
+public interface UserClient extends UserService {
 
     @GetMapping("/api/v1/user/master/{userId}")
-    UserResponse getUserByMaster(@PathVariable UUID userId);
+    UserResponse getUserByMaster(@PathVariable(value = "userId") UUID userId);
 }
