+++ deliverer/src/main/java/com/sparta/rooibos/deliverer/infrastructure/util/AuthorizationUtil.java
@@ -1,8 +1,8 @@
 package com.sparta.rooibos.deliverer.infrastructure.util;
 
 import com.querydsl.core.BooleanBuilder;
+import com.sparta.rooibos.deliverer.application.auditing.UserAuditorContext;
 import com.sparta.rooibos.deliverer.domain.entity.QDeliverer;
-import com.sparta.rooibos.deliverer.infrastructure.auditing.UserAuditorContext;
 import com.sparta.rooibos.deliverer.infrastructure.feign.HubClient;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
