+++ deliverer/src/main/java/com/sparta/rooibos/deliverer/presentation/controller/DelivererController.java
@@ -5,8 +5,9 @@ import com.sparta.rooibos.deliverer.application.dto.request.DelivererSearchReque
 import com.sparta.rooibos.deliverer.application.dto.response.DelivererListResponse;
 import com.sparta.rooibos.deliverer.application.dto.response.DelivererResponse;
 import com.sparta.rooibos.deliverer.application.service.port.DelivererService;
-import com.sparta.rooibos.deliverer.domain.entity.DelivererType;
-import com.sparta.rooibos.deliverer.infrastructure.aop.RoleCheck;
+
+import com.sparta.rooibos.deliverer.application.aop.RoleCheck;
+
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -57,7 +58,7 @@ public class DelivererController {
     @RoleCheck({"ROLE_MASTER, ROLE_HUB, ROLE_DELIBERY"})
     public ResponseEntity<DelivererResponse> assignDeliverer(
             @RequestParam UUID hubId,
-            @RequestParam DelivererType type
+            @RequestParam String type
     ) {
         DelivererResponse assigned = delivererService.assignNextDeliverer(hubId, type);
         return ResponseEntity.ok(assigned);
