+++ hub/src/main/java/com/sparta/rooibos/hub/application/core/HubManagerServiceImpl.java
@@ -1,5 +1,7 @@
 package com.sparta.rooibos.hub.application.core;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 import com.sparta.rooibos.hub.application.dto.hubManager.request.CreateHubManagerRequest;
 import com.sparta.rooibos.hub.application.dto.hubManager.response.CreateHubManagerResponse;
 import com.sparta.rooibos.hub.application.exception.BusinessHubManagerException;
@@ -7,8 +9,6 @@ import com.sparta.rooibos.hub.application.exception.custom.HubManagerErrorCode;
 import com.sparta.rooibos.hub.application.port.in.HubManagerService;
 import com.sparta.rooibos.hub.domain.model.HubManager;
 import com.sparta.rooibos.hub.domain.respository.HubManagerRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
 import java.util.UUID;
