+++ hub/src/main/java/com/sparta/rooibos/hub/infrastructure/client/NaverGeocodingClient.java
@@ -1,11 +1,11 @@
 package com.sparta.rooibos.hub.infrastructure.client;
 
-import com.sparta.rooibos.hub.application.dto.GeoLocation.response.GetCoordinatesResponse;
-import com.sparta.rooibos.hub.infrastructure.config.FeignClientRetryConfig;
-import com.sparta.rooibos.hub.infrastructure.config.NaverFeignClientConfig;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import com.sparta.rooibos.hub.application.dto.GeoLocation.response.GetCoordinatesResponse;
+import com.sparta.rooibos.hub.infrastructure.config.FeignClientRetryConfig;
+import com.sparta.rooibos.hub.infrastructure.config.NaverFeignClientConfig;
 
 @FeignClient(
         name = "naverGeocodingClient",
