+++ hub/src/main/java/com/sparta/rooibos/hub/infrastructure/jpa/JpaHubRepository.java
@@ -1,9 +1,9 @@
 package com.sparta.rooibos.hub.infrastructure.jpa;
 
-import com.sparta.rooibos.hub.domain.model.Hub;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+import com.sparta.rooibos.hub.domain.model.Hub;
 
 import java.util.Optional;
 import java.util.UUID;
