+++ hub/src/main/java/com/sparta/rooibos/hub/presentation/adapter/in/controller/HubController.java
@@ -1,5 +1,8 @@
 package com.sparta.rooibos.hub.presentation.adapter.in.controller;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
 import com.sparta.rooibos.hub.application.aop.RoleCheck;
 import com.sparta.rooibos.hub.application.dto.hub.request.CreateHubRequest;
 import com.sparta.rooibos.hub.application.dto.hub.request.SearchHubRequest;
@@ -9,9 +12,6 @@ import com.sparta.rooibos.hub.application.dto.hub.response.GetHubResponse;
 import com.sparta.rooibos.hub.application.dto.hub.response.SearchHubResponse;
 import com.sparta.rooibos.hub.application.dto.hub.response.UpdateHubResponse;
 import com.sparta.rooibos.hub.application.port.in.HubService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
 
 import java.util.UUID;
