+++ route/src/main/java/com/sparta/rooibos/route/infrastructure/client/HubFeignClient.java
@@ -1,11 +1,11 @@
 package com.sparta.rooibos.route.infrastructure.client;
 
-import com.sparta.rooibos.route.application.dto.response.hub.HubClientResponse;
-import com.sparta.rooibos.route.infrastructure.config.FeignClientRetryConfig;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
+import com.sparta.rooibos.route.application.dto.response.hub.HubClientResponse;
+import com.sparta.rooibos.route.infrastructure.config.FeignClientRetryConfig;
 
 import java.util.UUID;
