+++ route/src/main/java/com/sparta/rooibos/route/presentation/adapter/in/message/HubDeletedActivityListener.java
@@ -2,11 +2,11 @@ package com.sparta.rooibos.route.presentation.adapter.in.message;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sparta.rooibos.route.application.port.in.RouteService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.json.JsonParseException;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
+import com.sparta.rooibos.route.application.port.in.RouteService;
 
 @Component
 @RequiredArgsConstructor
